# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Compliance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compliance-items:
    get:
      operationId: listComplianceItems
      tags: [compliance-items]
      summary: List all compliance items
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, overdue]
        - name: siteId
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: contractorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of compliance items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceItem"
    post:
      operationId: createComplianceItem
      tags: [compliance-items]
      summary: Create a compliance item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplianceItemRequest"
      responses:
        "201":
          description: Created compliance item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
  /compliance-items/{id}:
    get:
      operationId: getComplianceItem
      tags: [compliance-items]
      summary: Get a compliance item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Compliance item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateComplianceItem
      tags: [compliance-items]
      summary: Update a compliance item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplianceItemRequest"
      responses:
        "200":
          description: Updated compliance item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteComplianceItem
      tags: [compliance-items]
      summary: Delete a compliance item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /compliance-items/{id}/status:
    patch:
      operationId: updateComplianceItemStatus
      tags: [compliance-items]
      summary: Update compliance item status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Updated compliance item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ComplianceItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
        priority:
          type: string
          enum: [low, medium, high, critical]
        siteId:
          type: integer
          nullable: true
        siteName:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        categoryColor:
          type: string
          nullable: true
        contractorId:
          type: integer
          nullable: true
        contractorName:
          type: string
          nullable: true
        contractorEmail:
          type: string
          nullable: true
        assignedTo:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        leadTimeDays:
          type: integer
          nullable: true
        notificationSentAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - createdAt
        - updatedAt
    CreateComplianceItemRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
          default: pending
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        siteId:
          type: integer
          nullable: true
        categoryId:
          type: integer
          nullable: true
        contractorId:
          type: integer
          nullable: true
        assignedTo:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        leadTimeDays:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - title
    UpdateComplianceItemRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
        priority:
          type: string
          enum: [low, medium, high, critical]
        siteId:
          type: integer
          nullable: true
        categoryId:
          type: integer
          nullable: true
        contractorId:
          type: integer
          nullable: true
        assignedTo:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        leadTimeDays:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
      required:
        - status
...